Change return type to Double for squareRoot

+++ cppsrc/square-root.h
@@ -2,7 +2,7 @@
 
 namespace mathlech
 {
-float squareRoot(int num, int precision);
+double squareRoot(int num, int precision);
 Napi::Number SquareRootWrapped(const Napi::CallbackInfo &info);
 Napi::Object Init(Napi::Env env, Napi::Object exports);
 } // namespace mathlech
